# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/signin": {
            "post": {
                "description": "Sign In using email and password if you have signed up before",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DELOS CRUD-APP"
                ],
                "summary": "Sign In",
                "parameters": [
                    {
                        "description": "Sign In Payload",
                        "name": "SignInRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SignInRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SignInResponse"
                        }
                    }
                }
            }
        },
        "/signup": {
            "post": {
                "description": "Sign Up using email and password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DELOS CRUD-APP"
                ],
                "summary": "Sign Up",
                "parameters": [
                    {
                        "description": "Sign Up Payload",
                        "name": "SignUpRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SignUpRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SignUpResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.SignInRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "risywanazriel@gmail.com"
                },
                "password": {
                    "type": "string",
                    "example": "123456"
                }
            }
        },
        "dto.SignInResponse": {
            "type": "object",
            "properties": {
                "user": {
                    "type": "object",
                    "properties": {
                        "access_token": {
                            "type": "string"
                        },
                        "expires_in": {
                            "type": "integer"
                        },
                        "provider_refresh_token": {
                            "type": "string"
                        },
                        "provider_token": {
                            "type": "string"
                        },
                        "refresh_token": {
                            "type": "string"
                        },
                        "token_type": {
                            "type": "string"
                        },
                        "user": {
                            "type": "object",
                            "properties": {
                                "app_metadata": {
                                    "type": "object",
                                    "additionalProperties": true
                                },
                                "aud": {
                                    "type": "string"
                                },
                                "confirmation_sent_at": {
                                    "type": "string"
                                },
                                "confirmed_at": {
                                    "type": "string"
                                },
                                "created_at": {
                                    "type": "string"
                                },
                                "email": {
                                    "type": "string"
                                },
                                "id": {
                                    "type": "string"
                                },
                                "invited_at": {
                                    "type": "string"
                                },
                                "role": {
                                    "type": "string"
                                },
                                "updated_at": {
                                    "type": "string"
                                },
                                "user_metadata": {
                                    "$ref": "#/definitions/dto.UserMetadata"
                                }
                            }
                        }
                    }
                }
            }
        },
        "dto.SignUpRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "dto.SignUpResponse": {
            "type": "object",
            "properties": {
                "user": {
                    "type": "object",
                    "properties": {
                        "app_metadata": {
                            "type": "object",
                            "additionalProperties": true
                        },
                        "aud": {
                            "type": "string"
                        },
                        "confirmation_sent_at": {
                            "type": "string"
                        },
                        "confirmed_at": {
                            "type": "string"
                        },
                        "created_at": {
                            "type": "string"
                        },
                        "email": {
                            "type": "string"
                        },
                        "id": {
                            "type": "string"
                        },
                        "invited_at": {
                            "type": "string"
                        },
                        "role": {
                            "type": "string"
                        },
                        "updated_at": {
                            "type": "string"
                        },
                        "user_metadata": {
                            "$ref": "#/definitions/dto.UserMetadata"
                        }
                    }
                }
            }
        },
        "dto.UserMetadata": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "email_verified": {
                    "type": "boolean"
                },
                "phone_verified": {
                    "type": "boolean"
                },
                "sub": {
                    "type": "string"
                }
            }
        }
    }
}